### 2018-02-18 #586 @caseyyoungflesh
Suggested reviewers (based on 'MCMC' or 'Bayes' in Domains and 'R' in Preferred Language):

* pboesu
* mdogucu
* jkarreth
* paulgovan
