### 2018-03-22 #586 @arfon
@jkarreth - would you be willing to review this submission for JOSS?
