### 2018-03-23 #586 @jkarreth
Yes, happy to review it!
