### 2018-03-23 #640 @whedon
[REVIEW]: MCMCvis: Tools to Visualize, Manipulate, and Summarize MCMC Output
**Submitting author:** @caseyyoungflesh (<a href="http://orcid.org/0000-0001-6343-3311">Casey Youngflesh</a>)
**Repository:** <a href="https://github.com/caseyyoungflesh/MCMCvis" target ="_blank">https://github.com/caseyyoungflesh/MCMCvis</a>
**Version:** v0.10.0
