### 2018-03-29 #640 @caseyyoungflesh
@jkarreth Many thanks. I'll respond here once these issues are addressed.

## Responses to major comments
* Thanks for pointing that bug out. I'll have it fixed shortly.
* 'tidybayes' appears to be an excellent package that I was not aware of until fairly recently. Its strengths lie in its integration of Bayesian model output with ggplot. I think the area where `MCMCvis` excels beyond what `tidybayes` has to offer is summarizing and manipulating Bayesian model output. As you say, the two packages both make analyzing Bayesian model output easier, though the emphasis differs. I can see the additional plotting functionality of `tidybayes` to be extremely useful. Combining these two packages would certainly make for a useful 'one stop shop' for dealing with Bayesian model output. I think there may be an opportunity for merging these packages in the future and I appreciate the suggestion.

## Responses to minor comments
* While all functions have examples, I agree that creating several toy models using different packages to show that the package can handle different output would be helpful for users. I will update the vignette accordingly.
* I think including some tests will make future development easier. I will include these with the other suggested improvements.
