### 2018-03-30 977f23a
fix warning with jags.parallel object
