### 2018-03-30 94a830f
Fix stan.fit object issue
