### 2018-03-30 40dc657
fix issue with jasgparallel
