+++ R/MCMCpstr.R
@@ -36,11 +36,6 @@ MCMCpstr <- function(object,
                    func = mean)
 {
   #SORTING BLOCK
-  if(class(object) == 'jagsUI')
-  {
-    object <- object$samples
-  }
-
   if(typeof(object) == 'double')
   {
     object2 <- MCMCchains(object, params, excl, ISB, mcmc.list = FALSE)
