### 2018-03-30 9e3279a
add pstr tests
