### 2018-03-30 3306cdd
add MCMCsummary tests
