### 2018-03-30 2fa819c
add ORCID id and testthat
