### 2018-03-30 46dbbd1
Add color arg and change ref_ovl def to F
