### 2018-03-30 4dcacdd
MCMCtrace - add gvals (generating values)
