+++ man/MCMCplot.Rd
@@ -5,10 +5,11 @@
 \title{Caterpillar plots of posterior distributions from MCMC output}
 \usage{
 MCMCplot(object, params = "all", excl = NULL, ISB = TRUE, ref = 0,
-  ref_ovl = TRUE, rank = FALSE, horiz = TRUE, xlim, ylim, xlab, ylab,
-  main, labels, labels_sz = 1.2, med_sz = 1.5, thick_sz = 5,
-  thin_sz = 2, ax_sz = 3, axis_text_sz = 1.3, tick_text_sz = 1.2,
-  main_text_sz = 1.2, tick_pos, mar = c(5.1, 4.1, 4.1, 2.1))
+  ref_ovl = FALSE, col = "black", rank = FALSE, horiz = TRUE, xlim,
+  ylim, xlab, ylab, main, labels, labels_sz = 1.2, med_sz = 1.5,
+  thick_sz = 5, thin_sz = 2, ax_sz = 3, axis_text_sz = 1.3,
+  tick_text_sz = 1.2, main_text_sz = 1.2, tick_pos, mar = c(5.1, 4.1, 4.1,
+  2.1))
 }
 \arguments{
 \item{object}{Object containing MCMC output. See DETAILS below.}
@@ -28,6 +29,8 @@ Default is \code{ref = 0}.
 Argument \code{NULL} will plot no reference line.}
 
 \item{ref_ovl}{Logical specifying whether the style/color of plotted median dots and CI should be changed based on whether the 50 \% and 95 \% credible intervals overlap the reference line. See DETAILS for more information.}
+
+\item{col}{Character string specifying which color to render estimates on plot. When \code{ref_ovl = TRUE}, this argument has no effect and colors plotted will be based on the credible intervals and reference line.}
 
 \item{rank}{Logical specifying whether output should be ranked. If \code{TRUE} posteriors will be ranked in decreasing order (based on specified measure of centrality) from top down.}
 
@@ -80,7 +83,7 @@ Visualize posterior distributions from MCMC output for specific parameters of in
 }
 \section{Details}{
 
-Points represent posterior medians. Parameters where 50\% credible intervals overlap 0 (or other specified value) are indicated by 'open' circles. Parameters where 50 percent credible intervals DO NOT overlap 0 AND 95 percent credible intervals DO overlap 0 (or other specified value) are indicated by 'closed' gray circles. Parameters where 95 percent credible intervals DO NOT overlap 0 (or other specified value) are indicated by 'closed' black circles. Thick lines represent 50 percent credible intervals while thin lines represent 95 \% credible intervals. \code{ref_ovl = FALSE} can be used to disable this feature. All median dots will be represented as 'closed' black circles.
+Points represent posterior medians. Parameters where 50\% credible intervals overlap 0 (or other specified value) are indicated by 'open' circles. Parameters where 50 percent credible intervals DO NOT overlap 0 AND 95 percent credible intervals DO overlap 0 (or other specified value) are indicated by 'closed' gray circles. Parameters where 95 percent credible intervals DO NOT overlap 0 (or other specified value) are indicated by 'closed' black circles. Thick lines represent 50 percent credible intervals while thin lines represent 95 \% credible intervals. \code{ref_ovl = TRUE} can be used to enable this feature.
 
 \code{object} argument can be a \code{stanfit} object (\code{rstan} package), an \code{mcmc.list} object (\code{coda} package), an \code{R2jags} model object (\code{R2jags} package), a \code{jagsUI} model object (\code{jagsUI} package), or a matrix containing MCMC chains (each column representing MCMC output for a single parameter, rows representing iterations in the chain). The function automatically detects the object type and proceeds accordingly.
 }
