+++ man/MCMCtrace.Rd
@@ -5,8 +5,8 @@
 \title{Trace and density plots from MCMC output}
 \usage{
 MCMCtrace(object, params = "all", excl = NULL, ISB = TRUE, iter = 5000,
-  priors = NULL, pdf = TRUE, filename, wd = getwd(), type = "both",
-  ind = FALSE)
+  gvals = NULL, priors = NULL, pdf = TRUE, open_pdf = TRUE, filename,
+  wd = getwd(), type = "both", ind = FALSE)
 }
 \arguments{
 \item{object}{Object containing MCMC output. See DETAILS below.}
@@ -21,9 +21,13 @@ Default \code{'all'} returns chains for all parameters.}
 
 \item{iter}{Number of iterations to plot for trace and density plots. The default value is 5000, meaning the last 5000 iterations of the chain will be plotted.}
 
+\item{gvals}{Vector containing generating values if simulated data was used to fit model. These values will be plotted as vertical lines on the density plots to compare posterior distributions with the true parameter values used to generate the data. No line will be apparent if the generating value is outside the plotted range of the posterior distribution.}
+
 \item{priors}{Matrix containing random draws from prior distributions corresponding to parameters of interest. If specified, priors are plotted along with posterior density plots. Percent overlap between prior and posterior is also calculated and displayed on each plot. Each column of the matrix represents a prior for a different parameter. Parameters are plotted alphabetically - priors should be sorted accordingly. If \code{priors} contains only one prior and more than one parameter is specified for the \code{params} argument, this prior will be used for all parameters. The number of draws for each prior should equal the number of iterations specified by \code{iter} (or total draws if less than \code{iter}) times the number of chains, though the function will automatically adjust if more or fewer iterations are specified. See DETAILS below.}
 
 \item{pdf}{Logical - if \code{pdf = TRUE} plots will be exported to a pdf.}
+
+\item{open_pdf}{Logical - if \code{open_pdf = TRUE} pdf will open in viewer after being generated.}
 
 \item{filename}{Name of pdf file to be printed. Default is 'MCMCtrace'.}
 
