+++ vignettes/MCMCvis.Rmd
@@ -23,17 +23,70 @@ The package contains five functions:
 
 `MCMCvis` was designed to perform key functions for MCMC analysis using minimal code, in order to free up time/brainpower for interpretation of analysis results. Functions support simple and straightforward subsetting of model parameters within the calls, and produce presentable and 'publication-ready' output.
 
+All functions in the package accept `stanfit` objects (created with the `rstan` package), `mcmc.list` objects (created with the `rjags` or `coda` packages), `R2jags` output (created with the `R2jags` package), `jagsUI` output (created with the `jagsUI` package), and matrices of MCMC output (one chain per column). The functions automatically detect the object type and proceed accordingly. Output objects can be inserted directly into the `MCMCvis` functions as an argument.
+
+#### JAGS model
+
+```{r, message= FALSE, warning = FALSE, results = FALSE}
+library(rjags)
+library(MCMCvis)
+
+#create JAGS model
+mf <- "
+model {
+for (i in 1:3)
+{
+  mu[i] ~ dnorm(0, 0.01);
+}
+}
+"
+
+jm <- jags.model(textConnection(mf),
+                 n.chains = 3)
+jags_out <- coda.samples(jm,
+                         variable.names = 'mu',
+                         n.iter = 100)
+```
+```{r}
+#plug object directly into package function
+MCMCsummary(jags_out)
+```
+
+#### Stan model
+
+```{r, message= FALSE, warning = FALSE, results = FALSE}
+library(rstan)
+
+#create Stan model
+sm <- "
+parameters {
+real mu[3]; 
+} 
+model {
+for (i in 1:3)
+{
+  mu[i] ~ normal(0, 10);
+}
+} 
+"
+
+stan_out <- stan(model_code = sm, iter = 50) 
+```
+```{r}
+#plug object directly into package function
+MCMCsummary(stan_out)
+```
+
 
 &nbsp;
 
-
 ## MCMCsummary
 
-`MCMCsummary` is used to output summary information from MCMC output. This function accepts `stanfit` objects (created with the `rstan` package), `mcmc.list` objects (created with the `rjags` or `coda` packages), `R2jags` output (created with the `R2jags` package), `jagsUI` output (created with the `jagsUI` package), and matrices of MCMC output (one chain per column). The function automatically detects the object type and proceeds accordingly. Two decimal places are reported by default. This can be changed using the `digits` argument.
+`MCMCsummary` is used to output summary information from MCMC output. Two decimal places are reported by default. This can be changed using the `digits` argument. We'll use the build in `mcmc.list` object data for the examples below, but model output of any of the supported types will behave in the same way.
+
+
 
 ```{r, message=FALSE}
-library(MCMCvis)
-
 data(MCMC_data)
 
 MCMCsummary(MCMC_data)
@@ -72,7 +125,7 @@ MCMCsummary(MCMC_data,
 
 &nbsp;
 
-Setting the `Rhat` and `n.eff` arguments to `FALSE` can be used to avoid calculating the Rhat statistic and number of effective samples, respectively (default for `Rhat` and `n.eff` are `TRUE` and `FALSE`, respectively). Specifying `FALSE` can greatly increase function speed with very large `mcmc.list` objects. Values for Rhat near 1 suggest convergence (Brooks and Gelman 1998). Kruschke (2014) recommends n.eff > 10,000 for reasonably stable posterior estimates.
[…]
