+++ man/MCMCsummary.Rd
@@ -69,8 +69,8 @@ MCMCsummary(MCMC_data)
 #Just 'beta' parameters
 MCMCsummary(MCMC_data, params = 'beta')
 
-#Just 'beta[1]', 'gamma[4]', and 'alpha[3]'
+#Just 'beta[1]', 'beta[4]', and 'alpha[3]'
 #'params' takes regular expressions when ISB = FALSE, square brackets must be escaped with '\\\\'
-MCMCsummary(MCMC_data, params = c('beta\\\\[1\\\\]', 'gamma\\\\[4\\\\]', 'alpha\\\\[3\\\\]'), ISB = FALSE)
+MCMCsummary(MCMC_data, params = c('beta\\\\[1\\\\]', 'beta\\\\[4\\\\]', 'alpha\\\\[3\\\\]'), ISB = FALSE)
 
 }
