### 2018-03-30 9680dfa
Fix examples
