+++ R/MCMCchains.R
@@ -31,9 +31,9 @@
 #' ex2 <- MCMCchains(MCMC_data, params = 'beta')
 #' apply(ex2, 2, mean)
 #'
-#' #Just 'beta[1]', 'gamma[4]', and 'alpha[3]'
+#' #Just 'beta[1]', 'beta[4]', and 'alpha[3]'
 #' #'params' takes regular expressions when ISB = FALSE, square brackets must be escaped with '\\'
-#' ex3 <- MCMCchains(MCMC_data, params = c('beta\\[1\\]', 'gamma\\[4\\]', 'alpha\\[3\\]'), ISB = FALSE)
+#' ex3 <- MCMCchains(MCMC_data, params = c('beta\\[1\\]', 'beta\\[4\\]', 'alpha\\[3\\]'), ISB = FALSE)
 #' apply(ex3, 2, sd)
 #'
 #' @export
