### 2018-03-30 dbcda25
update vignette
