+++ vignettes/MCMCvis.Rmd
@@ -23,13 +23,13 @@ The package contains five functions:
 
 `MCMCvis` was designed to perform key functions for MCMC analysis using minimal code, in order to free up time/brainpower for interpretation of analysis results. Functions support simple and straightforward subsetting of model parameters within the calls, and produce presentable and 'publication-ready' output.
 
-All functions in the package accept `stanfit` objects (created with the `rstan` package), `mcmc.list` objects (created with the `rjags` or `coda` packages), `R2jags` output (created with the `R2jags` package), `jagsUI` output (created with the `jagsUI` package), and matrices of MCMC output (one chain per column). The functions automatically detect the object type and proceed accordingly. Output objects can be inserted directly into the `MCMCvis` functions as an argument.
+All functions in the package accept `stanfit` objects (created with the `rstan` package), `mcmc.list` objects (created with the `rjags` or `coda` packages), `R2jags` output (created with the `R2jags` package), `jagsUI` output (created with the `jagsUI` package), and matrices of MCMC output (one chain per column - columns to be named with parameter names). The functions automatically detect the object type and proceed accordingly. Output objects can be inserted directly into the `MCMCvis` functions as an argument.
 
 #### JAGS model
 
 ```{r, message= FALSE, warning = FALSE, results = FALSE}
-library(rjags)
 library(MCMCvis)
+library(rjags)
 
 #create JAGS model
 mf <- "
@@ -41,9 +41,9 @@ for (i in 1:3)
 }
 "
 
-jm <- jags.model(textConnection(mf),
+jm <- rjags::jags.model(textConnection(mf),
                  n.chains = 3)
-jags_out <- coda.samples(jm,
+jags_out <- rjags::coda.samples(jm,
                          variable.names = 'mu',
                          n.iter = 100)
 ```
@@ -70,7 +70,7 @@ for (i in 1:3)
 } 
 "
 
-stan_out <- stan(model_code = sm, iter = 50) 
+stan_out <- rstan::stan(model_code = sm, iter = 10) 
 ```
 ```{r}
 #plug object directly into package function
@@ -334,9 +334,7 @@ MCMCplot(MCMC_data,
        xlab = 'My x-axis label',
        main = 'MCMCvis plot',
        labels = c('First param', 'Second param', 'Third param', 
-                  'Fourth param', 'Fifth param', 'Sixth param', 
-                  'Seventh param', 'Eighth param', 'Nineth param', 
-                  'Tenth param'), 
+                  'Fourth param', 'Fifth param', 'Sixth param'), 
        col = 'red',
        labels_sz = 1.5,
        med_sz = 2,
