### 2018-03-31 056e811
remove need for jags packages
