### 2018-03-31 908a1ab
remove need for jags install frmo yml file
