### 2018-03-31 1b340b5
do not required jags and stan packages
