### 2018-03-31 c2b3c22
update CRAN comments
