+++ cran-comments.md
@@ -1,34 +1,17 @@
 CRAN comments
 ====
 
-## Resubmission
-This is a resubmission. As of version 0.9.2, this package:
-
-* No longer opens external programs when running checks
-* Has depends set to patch-level 0
-
 
 ## Test environments
 * local OS X install, R 3.4.3
-* Ubuntu 14.04.5 (on travis-ci), R 3.4.2
+* Ubuntu 14.04.5 (on travis-ci), R 3.4.4
 * win-builder (devel)
 
 
 ## R CMD check results
-There were no ERRORs or WARNINGs.
 
-There was 1 NOTE:
+There were no ERRORs, WARNINGs, or NOTEs.
 
-'New submission
-
-Package was archived on CRAN
-
-CRAN repository db overrides:
-  X-CRAN-Comment: Archived on 2018-01-25 for policy violation (leaving
-    processes running).'
-    
-* This package no longer opens external processes.
-    
 
 ## Downstream dependencies
 I have run R CMD check on downstream dependencies of MCMCvis. All packages passed.
