### 2018-03-31 #640 @caseyyoungflesh
@jkarreth

Package version 0.10.0 is now on github.
*  This version fixes the `stanfit` bug identified above. `stanfit` objects are now processed correctly.
*  I have included toy models fit with JAGS and Stan in the vignette to demonstrate how output from different samplers can be input into the MCMCvis functions directly.
* I have also included automated tests to make sure that different input types are properly processed (to avoid the previously noted `stanfit` issue).
* This version also includes several other features, as noted in `NEWS.md`

MCMCvis version 0.10.0 has passed all CRAN checks. A new version will be submitted to CRAN soon. The most recent version github version can be installed with:

`devtools::install_github('caseyyoungflesh/MCMCvis', build_vignettes = TRUE)`
