### 2018-04-07 #640 @caseyyoungflesh
Thank you very much @jkarreth! Version 0.10.0 is now on CRAN (though all binaries may not be built yet).
