### 2018-04-10 #640 @caseyyoungflesh
@arfon thanks!

**DOI for all versions:** 10.5281/zenodo.1216119

**DOI for v0.10.0:** 10.5281/zenodo.1216120
